+++ tests/test_scene.py
@@ -31,10 +31,10 @@ class TestScene:
         s_or_fp = transform(geojson_file)
         scene = Scene.from_geojson(s_or_fp, tx_loc="SW", rx_loc="NE")
         bounding_box = scene.bounding_box()
-        assert len(scene.emitters) == 1
+        assert len(scene.transmitters) == 1
         assert len(scene.objects) == 28
         assert len(scene.receivers) == 1
-        chex.assert_trees_all_equal(scene.emitters["tx"].point, bounding_box[0, :])
+        chex.assert_trees_all_equal(scene.transmitters["tx"].point, bounding_box[0, :])
         chex.assert_trees_all_equal(scene.receivers["rx"].point, bounding_box[1, :])
 
     def test_from_geojson_unimplemented(self, geojson_file):
@@ -62,24 +62,24 @@ class TestScene:
         ],
     )
     def test_random_uniform_scene(self, key, n):
-        scene = Scene.random_uniform_scene(key, n_emitters=n)
+        scene = Scene.random_uniform_scene(key, n_transmitters=n)
 
         assert isinstance(scene, Scene)
-        assert len(scene.emitters) == n
+        assert len(scene.transmitters) == n
         assert len(scene.objects) == 1
         assert len(scene.receivers) == 1
 
         scene = Scene.random_uniform_scene(key, n_walls=n)
 
         assert isinstance(scene, Scene)
-        assert len(scene.emitters) == 1
+        assert len(scene.transmitters) == 1
         assert len(scene.objects) == n
         assert len(scene.receivers) == 1
 
         scene = Scene.random_uniform_scene(key, n_receivers=n)
 
         assert isinstance(scene, Scene)
-        assert len(scene.emitters) == 1
+        assert len(scene.transmitters) == 1
         assert len(scene.objects) == 1
         assert len(scene.receivers) == n
 
@@ -107,7 +107,7 @@ class TestScene:
         assert len(scene.objects) == 8
 
     def test_plot(self, ax, key):
-        scene = Scene.random_uniform_scene(key, n_emitters=3, n_walls=5, n_receivers=2)
+        scene = Scene.random_uniform_scene(key, n_transmitters=3, n_walls=5, n_receivers=2)
         _ = scene.plot(ax)
 
         scene = Scene.basic_scene()
@@ -123,7 +123,7 @@ class TestScene:
         scene = Scene.random_uniform_scene(key, n_walls=10)
 
         points = jnp.vstack(
-            [scene.emitters["tx_0"].point, scene.receivers["rx_0"].point]
+            [scene.transmitters["tx_0"].point, scene.receivers["rx_0"].point]
             + [obj.points for obj in scene.objects]
         )
 
@@ -137,12 +137,12 @@ class TestScene:
         chex.assert_trees_all_equal(expected, got)
         chex.assert_shape(got, (2, 2))
 
-    def test_get_closest_emitter(self, key):
-        scene = Scene.random_uniform_scene(key, n_emitters=10)
+    def test_get_closest_transmitter(self, key):
+        scene = Scene.random_uniform_scene(key, n_transmitters=10)
         expected_point = Point(point=jnp.array([0.5, 0.5]))
         expected_distance = jnp.array(0.0)
-        scene.emitters["closest"] = expected_point
-        got_point, got_distance = scene.get_closest_emitter(expected_point.point)
+        scene.transmitters["closest"] = expected_point
+        got_point, got_distance = scene.get_closest_transmitter(expected_point.point)
         chex.assert_trees_all_equal(got_point, expected_point)
         chex.assert_trees_all_equal(got_distance, expected_distance)
 
@@ -158,17 +158,17 @@ class TestScene:
     def test_accumulate_over_paths(self):
         pass
 
-    def test_accumulate_on_emitters_grid_over_paths(self):
-        def fun(emitter, receiver, path, interacting_objects):
+    def test_accumulate_on_transmitters_grid_over_paths(self):
+        def fun(transmitter, receiver, path, interacting_objects):
             return path.length() ** 2  # = x^2 + y^2 in LOS
 
         rx0 = Point(point=jnp.array([0.0, 0.0]))
         rx1 = Point(point=jnp.array([0.0, 1.0]))
-        scene = Scene(emitters={}, objects=[], receivers=dict(rx0=rx0, rx1=rx1))
+        scene = Scene(transmitters={}, objects=[], receivers=dict(rx0=rx0, rx1=rx1))
 
         x = y = jnp.linspace(-3, 3, 10)
         X, Y = jnp.meshgrid(x, y)
-        got_Z = scene.accumulate_on_emitters_grid_over_paths(
+        got_Z = scene.accumulate_on_transmitters_grid_over_paths(
             X, Y, fun=fun, max_order=1, approx=False
         )
 
@@ -187,13 +187,13 @@ class TestScene:
         expected_Z1 = X**2 + (Y - 1.0) ** 2
         chex.assert_trees_all_close(got_Z1, expected_Z1)
 
-        got_Z = scene.accumulate_on_emitters_grid_over_paths(
+        got_Z = scene.accumulate_on_transmitters_grid_over_paths(
             X, Y, fun=fun, reduce_all=True, max_order=1, approx=False
         )
         expected_Z = expected_Z0 + expected_Z1
         chex.assert_trees_all_close(got_Z, expected_Z)
 
-        got_dZ = scene.accumulate_on_emitters_grid_over_paths(
+        got_dZ = scene.accumulate_on_transmitters_grid_over_paths(
             X, Y, fun=fun, reduce_all=True, grad=True, max_order=1, approx=False
         )
         expected_dZ0 = jnp.stack([2 * X, 2 * Y], axis=-1)
@@ -201,7 +201,7 @@ class TestScene:
         expected_dZ = expected_dZ0 + expected_dZ1
         chex.assert_trees_all_close(got_dZ, expected_dZ)
 
-        got_Z, got_dZ = scene.accumulate_on_emitters_grid_over_paths(
+        got_Z, got_dZ = scene.accumulate_on_transmitters_grid_over_paths(
             X,
             Y,
             fun=fun,
@@ -214,12 +214,12 @@ class TestScene:
         chex.assert_trees_all_close(got_dZ, expected_dZ)
 
     def test_accumulate_on_receivers_grid_over_paths(self):
-        def fun(emitter, receiver, path, interacting_objects):
+        def fun(transmitter, receiver, path, interacting_objects):
             return path.length() ** 2  # = x^2 + y^2 in LOS
 
         tx0 = Point(point=jnp.array([0.0, 0.0]))
         tx1 = Point(point=jnp.array([1.0, 0.0]))
-        scene = Scene(emitters=dict(tx0=tx0, tx1=tx1), objects=[], receivers={})
+        scene = Scene(transmitters=dict(tx0=tx0, tx1=tx1), objects=[], receivers={})
 
         x = y = jnp.linspace(-3, 3, 10)
         X, Y = jnp.meshgrid(x, y)